requirements.txt
selenium
discord.py